refactor: SelectionOption 리팩터링
- 패키지 위치 변경
- 개행 변경
